add jcommander, fix interest calculation

+++ src/main/java/by/clevertec/model/DiscountCard.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -12,5 +14,5 @@ import lombok.NoArgsConstructor;
 public class DiscountCard {
     private int id;
     private String ownersEmail;
-    private int discount;
+    private BigDecimal discount;
 }
